[be/refactor] Integer -> Long 자료형 변경

See merge request s08-bigdata-dist-sub2/S08P22A202!218

+++ backend/bid/src/main/java/com/dokidoki/bid/api/request/AuctionBidReq.java
@@ -14,6 +14,6 @@ import lombok.ToString;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class AuctionBidReq {
     private String name;
-    private int currentHighestPrice;
-    private int currentPriceSize;
+    private Long currentHighestPrice;
+    private Long currentPriceSize;
 }
